Move tests to right place.
